requirements.txt
# Dependencies

requests
python-dotenv
pandas
openpyxl